ref(hono): Consolidate route patching and add clarification comments
